Improved SearchService class by changing query from NxM down to 2

+++ src/main/java/com/example/hotelbookingapp/dto/CreateRoomRequest.java
@@ -1,10 +1,12 @@
 package com.example.hotelbookingapp.dto;
 
+import java.math.BigDecimal;
+
 public record CreateRoomRequest(
 
     Long hotelId,
     String roomNumber,
     String roomType,
-    double roomPrice
+    BigDecimal roomPrice
 
 ) {}
