+++ src/main/java/com/example/hotelbookingapp/dto/HotelSearchResponse.java
@@ -1,9 +1,11 @@
 package com.example.hotelbookingapp.dto;
 
+import java.math.BigDecimal;
+
 public record HotelSearchResponse(
     Long id,
     String name,
     String location,
     int availableRooms,
-    double minPrice
+    BigDecimal minPrice
 ) {}
